update: accessibily location and design
