+++ src/app/app.component.html
@@ -1,1 +1,2 @@
 <app-navbar /> <router-outlet />
+<app-accessibility />
