+++ src/app/components/accessibility/accessibility.component.html
@@ -1,0 +1,14 @@
+<div
+  class="invisible sm:visible fixed top-20 right-10 z-50 flex flex-col gap-3"
+>
+  <button
+    class="outline-none bg-primary-500 p-2 rounded-2xl opacity-30 hover:opacity-100 transition-opacity duration-300"
+  >
+    <svg class="size-7 text-primary-50" globe-icon></svg>
+  </button>
+  <button
+    class="outline-none bg-primary-500 p-2 rounded-2xl opacity-30 hover:opacity-100 transition-opacity duration-300"
+  >
+    <svg class="size-7 text-primary-50" moon-icon></svg>
+  </button>
+</div>
