--- src/app/features/landing/components/accessibility/accessibility.component.html
@@ -1,4 +1,0 @@
-<div class="flex items-center justify-center gap-12">
-  <svg class="size-7 text-primary-500" globe-icon></svg>
-  <svg class="size-7 text-primary-500" moon-icon></svg>
-</div>
